+++ src/cljs/cxengage_javascript_sdk/modules/sqs.cljs
@@ -1,6 +1,6 @@
 (ns cxengage-javascript-sdk.modules.sqs
   (:require-macros [cljs.core.async.macros :refer [go go-loop]]
-                   [lumbajack.macros :refer [log]])
+                   [cxengage-javascript-sdk.domain.macros :refer [log]])
   (:require [cxengage-javascript-sdk.domain.protocols :as pr]
             [cxengage-javascript-sdk.state :as state]
             [cljs.core.async :as a]